Made a selection group
Made a GroupSingle in pygame which contains up to one unit for the purposes of Ogre Battle-like unit selection. Updated the methods in the Unit class to add and remove themselves from the "selected" group. This should make it fairly easy to issue commands to units once I figure out how to make multiple distinct units.