Units stop when they reach their destination
Currently, only one destination can exist at a time

This results in bugs! Stuff like our units vibrating at their destination because there is no longer an actual destination to collide with, since the "pathfinding" is based on moving in a direction until collision with the destination happens. I even got a crash that I think involved normalizing a vector with length 0?? IDK

No such bugs happen if you only issue move commands to units and wait for them to arrive, which kills the destination and stops any unit behavior that relies on the destination, at which point a new move command on any other unit can be issued.

I did discover an old bug with the "pathfinding" though: (it's been here, but it became obvious while implementing collision) if the direction is set only once and doesn't update, then the unit can entirely miss the destination's collision rectangle!!! I'm assuming this has to do with rounding errors (the direction set was not accurate due to rounding errors in pygame's normalize method perhaps), but the solution I implemented was just to update the direction every frame while "pathfinding".