Made the beginnings of the ownership mechanic
In RTS games, player 1 has "ownership" of a certain set of selectables, and player 2 has "ownership" of a different set of selectables. The idea I had to do this was to go with a class attribute that indicates which player owns the unit. If I implement something like archon mode, this could be a list of owners.